pyproject.toml
[project]
name = "rate_owl"
version = "0.1.0"
description = "Rating and review serivice based on FastApi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.13",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "types-passlib>=1.7.7.20250602",
]

[tool.mypy]
exclude = [
    ".venv/",
    "migrations/versions"
]
